Start connecting examples to base profiles
Bundle examples have both Observation and MedicationAdministration resources, but we can tie them to only one profile explicitly. Which one?

Kanta PHR Patient example does not actually conform to IPA yet, as IPA still requires the name to be present. But the infrastructure does not check this, as the example is a JSON resource, not a FSH file...