Move diagnosis and reason for visit to new profile
Creates a "reason for care" profile. Moves diagnosis and käyntisyy stuff
there.

Overview of changes:
- Move all condition intro markdown content to reason-for-care intro
- Almost completely rewrites Scope and Usage from both intros
- Remove FiBaseConditionCategory as it's not needed anymore
-  Change rules for how encounter-diagnosis category is interpreted

Alters naming and wording so that "reason for visit" refers to THL
käyntisyy and the more general "reason for care" to both diagnosis
and käyntisyy. This should align concepts very closely to the THL's
Tiedonhallintapalvelu -spec.