Add an alternative Medication profile
Because the IPA profile for medications has constraints that are suitable for patient access use case, but not for all use cases.
This is just one potential approach. We could also use the more fine-grained mechanisms like the Obligations extension or other extensions for indicating conformance to another IG or profile.