Fix hetu identifier coding.
This is what we really want, right?

The validator still complains about the code `NNFIN`, but I'd say that's the problem of the validator. The code is formally presented as NNXXX in the value set, and verbally explained that you need to replace the X's with the three letter country code.