Move the dragon to the end of page
Now that the TOC is populated, it overlaps with the dragon note in a way that looks bad.
Not sure whether that's good enough for a reason to move the warning to the end of page (that's why this is in a separate commit, easy to roll back), but let's at least see.