Remove redundant agent references
I did not find a good way how to explain why to have these in the examples. The original idea was to illustrate how to make it more clear which agent the registry is relevant to, in case there would be more agents present in the same Provenance instance. This was discussed briefly with Kela, but I did not have good enough of a concept in my mind to be able to explain it.