Move Kanta examples to examples folder
This includes the KantaDocumentMetadata profile too. Not sure how we could tag that as an example and still reference it. I think an Instance might work, will try that next.

Add another example, to clarify role codings.
Also, changed the slicing so that the Kanta specific role is not always required.