Allow extensions from other ig's
Allow extensions from the IG's we know. This gets rid of many errors in the QA report.

An alternative would be to import also those extensions into this implementation guide. Then the use of the extensions would be controlled better too.
But it's not really our task to validate (and fix?) all those imported examples.