Remove register info
Remove the register type and register specifier info from Provenance.entity, as we weren't able to reach consensus on whether it should be there or somewhere else.
Still leave the removed parts visible in comments, so that it is easier to refer to them when the work continues in the future.