Tidy up example Kanta profile
Remove the extra imposes statement. The profile already inherits the Fi Base Information Cusstodian profile.

Also remove some commented code. This was left in the examples to wait for resolution on where and how tp store the register type and register specifier. However, in these examples the commented code may be more of a distraction. The comments that are left highlight what's the missing data.